Refactor room-related components to use course terminology
