+++ src/components/molecules/CourseModal.vue
@@ -20,7 +20,7 @@
 			</div>
 		</template>
 		<template #content>
-			<room-avatar-iterator
+			<course-avatar-iterator
 				class="iterator"
 				:avatars="groupData.groupElements"
 				:item-size="itemSize"
@@ -33,7 +33,7 @@
 	</vCustomDialog>
 </template>
 <script>
-import RoomAvatarIterator from "@/components/organisms/RoomAvatarIterator.vue";
+import CourseAvatarIterator from "@/components/organisms/CourseAvatarIterator.vue";
 import vCustomDialog from "@/components/organisms/vCustomDialog.vue";
 import { coursesModule } from "@/store";
 import { mdiKeyboardReturn, mdiPencilOutline } from "@mdi/js";
@@ -42,7 +42,7 @@ import { defineComponent } from "vue";
 export default defineComponent({
 	components: {
 		vCustomDialog,
-		RoomAvatarIterator,
+		CourseAvatarIterator,
 	},
 	props: {
 		isOpen: {
