+++ src/components/organisms/CourseAvatarIterator.vue
